feat: replace proxy implementations for storage helper
