--- zae_labeler/lib/src/utils/proxy_storage_helper/interface_storage_helper.dart
@@ -1,11 +1,0 @@
-import '../../models/project_model.dart';
-import '../../models/label_entry.dart';
-
-abstract class PlatformStorageHelper {
-  Future<List<Project>> loadProjects();
-  Future<void> saveProjects(List<Project> projects);
-  Future<List<LabelEntry>> loadLabelEntries();
-  Future<void> saveLabelEntries(List<LabelEntry> labelEntries);
-  Future<String> downloadLabelsAsZip(
-      Project project, List<LabelEntry> labelEntries, List<dynamic> dataFiles);
-}
